+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/dao/preset/AccountPresetCriteriaMaker.java
@@ -21,10 +21,15 @@ public class AccountPresetCriteriaMaker implements PresetCriteriaMaker {
 
     @Override
     public void makeCriteria(DetachedCriteria criteria, String preset, Object[] objs) {
-        if (AccountMaintainService.ID_LIKE.equals(preset)) {
-            idLike(criteria, objs);
-        } else {
-            throw new IllegalArgumentException("无法识别的预设: " + preset);
+        switch (preset) {
+            case AccountMaintainService.ID_LIKE:
+                idLike(criteria, objs);
+                break;
+            case AccountMaintainService.DISPLAY_NAME_LIKE:
+                displayNameLike(criteria, objs);
+                break;
+            default:
+                throw new IllegalArgumentException("无法识别的预设: " + preset);
         }
     }
 
@@ -33,6 +38,16 @@ public class AccountPresetCriteriaMaker implements PresetCriteriaMaker {
             String id = (String) objs[0];
             criteria.add(Restrictions.like("stringId", id, MatchMode.ANYWHERE));
             criteria.addOrder(Order.asc("stringId"));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objs));
+        }
+    }
+
+    private void displayNameLike(DetachedCriteria criteria, Object[] objs) {
+        try {
+            String id = (String) objs[0];
+            criteria.add(Restrictions.like("displayName", id, MatchMode.ANYWHERE));
+            criteria.addOrder(Order.asc("displayName"));
         } catch (Exception e) {
             throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objs));
         }